… experiments/benchmark_cairo_old.nim
… experiments/rounded_rect.nim
… experiments/test_svg_cairo.nim