before:
```
[pixie] rect opaque ................ 0.035 ms      0.036 ms    ±0.001  x1000
[pixie] rect not opaque ............ 0.120 ms      0.122 ms    ±0.003  x1000
[pixie] roundedRect opaque ......... 0.109 ms      0.113 ms    ±0.004  x1000
[pixie] roundedRect not opaque ..... 0.300 ms      0.303 ms    ±0.005  x1000
[pixie] pentagon opaque ............ 0.192 ms      0.196 ms    ±0.004  x1000
[pixie] pentagon not opaque ........ 0.378 ms      0.383 ms    ±0.006  x1000
[pixie] circle opaque .............. 0.243 ms      0.246 ms    ±0.004  x1000
[pixie] circle not opaque .......... 0.486 ms      0.497 ms    ±0.025  x1000
[pixie] heart opaque ............... 0.251 ms      0.254 ms    ±0.003  x1000
[pixie] heart not opaque ........... 0.468 ms      0.472 ms    ±0.006  x1000
[pixie] tiger ..................... 10.006 ms     10.120 ms    ±0.132   x493
```

after:
```
[pixie] rect opaque ................ 0.035 ms      0.037 ms    ±0.002  x1000 →
[pixie] rect not opaque ............ 0.123 ms      0.124 ms    ±0.003  x1000 →
[pixie] roundedRect opaque ......... 0.064 ms      0.067 ms    ±0.003  x1000 ↑
[pixie] roundedRect not opaque ..... 0.201 ms      0.205 ms    ±0.004  x1000 ↑
[pixie] pentagon opaque ............ 0.069 ms      0.072 ms    ±0.005  x1000 ↑
[pixie] pentagon not opaque ........ 0.129 ms      0.131 ms    ±0.002  x1000 ↑
[pixie] circle opaque .............. 0.108 ms      0.122 ms    ±0.023  x1000 ↑
[pixie] circle not opaque .......... 0.194 ms      0.197 ms    ±0.005  x1000 ↑
[pixie] heart opaque ............... 0.146 ms      0.149 ms    ±0.005  x1000 ↑
[pixie] heart not opaque ........... 0.264 ms      0.268 ms    ±0.007  x1000 ↑
[pixie] tiger ...................... 9.817 ms     10.053 ms    ±0.207   x496 ↑
```

cairo:
```
[cairo] rect opaque ................ 0.035 ms      0.037 ms    ±0.003  x1000 < tied
[cairo] rect not opaque ............ 0.142 ms      0.145 ms    ±0.008  x1000 < ! pixie faster
[cairo] roundedRect opaque ......... 0.071 ms      0.077 ms    ±0.006  x1000 < ! pixie faster
[cairo] roundedRect not opaque ..... 0.230 ms      0.231 ms    ±0.003  x1000 < ! pixie faster
[cairo] pentagon opaque ............ 0.093 ms      0.096 ms    ±0.005  x1000 < ! pixie faster
[cairo] pentagon not opaque ........ 0.245 ms      0.251 ms    ±0.014  x1000 < ! pixie faster
[cairo] circle opaque .............. 0.125 ms      0.128 ms    ±0.005  x1000 < ! pixie faster
[cairo] circle not opaque .......... 0.308 ms      0.321 ms    ±0.021  x1000 < ! pixie faster
[cairo] heart opaque ............... 0.142 ms      0.147 ms    ±0.007  x1000 < tied
[cairo] heart not opaque ........... 0.338 ms      0.344 ms    ±0.005  x1000 < ! pixie faster
[cairo] tiger ..................... 12.813 ms     12.990 ms    ±0.179   x384 < ! pixie faster
```